webnn/unary.html

<!doctype html>
<meta charset=utf-8>
<title>test element-wise unary operations</title>
<link rel="help" href="https://webmachinelearning.github.io/webnn/#api-mlgraphbuilder-unary">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="https://webmachinelearning.github.io/webnn-polyfill/dist/webnn-polyfill.js"></script>
<div id=log></div>
<script type="module">
  'use strict';
  import * as utils from './utils.js';
  let builder;
  setup(() => {
    const context = navigator.ml.createContext();
    builder = new MLGraphBuilder(context);
  })

  function testUnary(op, input, expected, shape) {
    const x = builder.input('x', {type: 'float32', dimensions: shape});
    const y = builder[op](x);
    const graph = builder.build({y});
    const inputs = {'x': new Float32Array(input)};
    const outputs = {'y': new Float32Array(utils.sizeOfShape(shape))};
    graph.compute(inputs, outputs);
    assert_array_approx_equals_ulp(outputs.y, expected, 0);
  }

  test(() => {
    testUnary('abs', [-1, 0, 1], [1, 0, 1], [3]);
    testUnary(
        'abs',
        [-1.1, 0, 1.1, 2.2, 0, -2.2],
        [1.1, 0, 1.1, 2.2, 0, 2.2],
        [2, 3]);
    testUnary(
        'abs',
        [-1.1, 0, 1.1, 2.2, 0, -2.2],
        [1.1, 0, 1.1, 2.2, 0, 2.2],
        [1, 2, 3]);
    testUnary(
        'abs',
        [-1.1, 0, 1.1, 2.2, 0, -2.2],
        [1.1, 0, 1.1, 2.2, 0, 2.2],
        [1, 2, 3, 1]);
  }, 'abs');

  test(() => {
    testUnary('ceil', [-1.1, 0, 1.1], [-1, 0, 2], [3]);
    testUnary(
        'ceil',
        [-1.1, 0, 1.1, -2.2, 0, 2.2],
        [-1, 0, 2, -2, 0, 3],
        [2, 3]);
    testUnary(
        'ceil',
        [-1.1, 0, 1.1, -2.2, 0, 2.2],
        [-1, 0, 2, -2, 0, 3],
        [1, 2, 3]);
    testUnary(
        'ceil',
        [-1.1, 0, 1.1, -2.2, 0, 2.2],
        [-1, 0, 2, -2, 0, 3],
        [1, 2, 3, 1]);
  }, 'ceil');

  test(() => {
    testUnary(
        'cos',
        [1.4124068, 1.9740626, -0.06506752, 0.73539704],
        [
          0.15772809760857773,
          -0.39242469654349826,
          0.9978838556864368,
          0.7415644450136674,
        ],
        [4]);
    testUnary(
        'cos',
        [
          1.4124068,   1.9740626,  -0.06506752, 0.73539704,
          -0.56439203, 0.89806247, 0.12939146,  -0.34816208,
          -1.0759926,  0.66291636, 0.21504708,  -0.71527237,
        ],
        [
          0.15772809760857773,
          -0.39242469654349826,
          0.9978838556864368,
          0.7415644450136674,
          0.8449139610653698,
          0.6231265199397442,
          0.9916405976730124,
          0.9400013446543031,
          0.4748589278431268,
          0.7882008050685133,
          0.9769663487271947,
          0.7549146426895217,
        ],
        [3, 4]);
    testUnary(
        'cos',
        [
          1.4124068,   1.9740626,
          -0.06506752, 0.73539704,
          -0.56439203, 0.89806247,
          0.12939146,  -0.34816208,
          -1.0759926,  0.66291636,
          0.21504708,  -0.71527237,
        ],
        [
          0.15772809760857773,
          -0.39242469654349826,
          0.9978838556864368,
          0.7415644450136674,
          0.8449139610653698,
          0.6231265199397442,
          0.9916405976730124,
          0.9400013446543031,
          0.4748589278431268,
          0.7882008050685133,
          0.9769663487271947,
          0.7549146426895217,
        ],
        [3, 2, 2]);
    testUnary(
        'cos',
        [
          1.4124068,
          1.9740626,
          -0.06506752,
          0.73539704,
          -0.56439203,
          0.89806247,
          0.12939146,
          -0.34816208,
          -1.0759926,
          0.66291636,
          0.21504708,
          -0.71527237,
        ],
        [
          0.15772809760857773,
          -0.39242469654349826,
          0.9978838556864368,
          0.7415644450136674,
          0.8449139610653698,
          0.6231265199397442,
          0.9916405976730124,
          0.9400013446543031,
          0.4748589278431268,
          0.7882008050685133,
          0.9769663487271947,
          0.7549146426895217,
        ],
        [3, 2, 2, 1]);
  }, 'cos');

  test(() => {
    testUnary('exp', [-1, 0, 1], [0.36787944117144233, 1, 2.718281828459045], [3]);
    testUnary(
        'exp',
        [
          0.3143407, 0.03632548, 0.5354084,  -0.5000897,
          1.2028517, -1.2581364, -1.5108215, -1.2340564,
          1.3860914, -0.2944251, -1.5065757, -0.4673513,
        ],
        [
          1.3693561967375985,
          1.036993312152022,
          1.708145706528859,
          0.6064762563524844,
          3.3295984129224556,
          0.2841831370156004,
          0.22072857493397907,
          0.29110932356722374,
          3.999188237901296,
          0.7449597417366962,
          0.2216677366530061,
          0.6266599061142515,
        ],
        [3, 4]);
    testUnary(
        'exp',
        [
          0.3143407,   0.03632548,  0.5354084,   -0.5000897,  1.2028517,
          -1.2581364,  -1.5108215,  -1.2340564,  1.3860914,   -0.2944251,
          -1.5065757,  -0.4673513,  0.56616277,  0.77866685,  -0.01097398,
          1.0758846,   0.6035437,   0.36806744,  0.03906458,  -0.54385495,
          0.10609569,  -0.40644982, -1.2890846,  1.3825086,   0.51489764,
          1.6407244,   -0.67886734, -0.6556329,  1.0399923,   0.1484657,
          1.011217,    0.8451463,   0.75473833,  -2.0161264,  1.6406634,
          -0.01692923, -0.7986609,  0.97758174,  0.893054,    -0.01632686,
          -1.9721986,  -0.75843745, 0.42327842,  -0.08648382, -1.3960054,
          0.7547995,   -0.42002508, -1.784105,   1.0171342,   0.3634587,
          0.4158588,   -1.0103701,  -0.23202766, 0.6390487,   -0.22796124,
          0.11259284,  0.3690759,   -0.18703128, 0.07711394,  2.9116163,
        ],
        [
          1.3693561967375985,
          1.036993312152022,
          1.708145706528859,
          0.6064762563524844,
          3.3295984129224556,
          0.2841831370156004,
          0.22072857493397907,
          0.29110932356722374,
          3.999188237901296,
          0.7449597417366962,
          0.2216677366530061,
          0.6266599061142515,
          1.7614948056979465,
          2.1785659734395244,
          0.9890860144586422,
          2.9325859189764807,
          1.828587297220383,
          1.4449394824067432,
          1.0398376341962599,
          0.580506111445723,
          1.111928271832297,
          0.666010515185213,
          0.27552288133235436,
          3.984885583357506,
          1.6734671969280184,
          5.158905269957428,
          0.5071911422650927,
          0.5191134117181747,
          2.829195229464421,
          1.1600530072738744,
          2.7489444455169916,
          2.328318422639379,
          2.127054864081674,
          0.13317031580877914,
          5.158590586333908,
          0.9832132641755142,
          0.4499310635787923,
          2.6580206785468157,
          2.4425779049391343,
          0.983805700764556,
          0.13915058318029658,
          0.4683977504013186,
          1.526959372817342,
          0.9171503881579351,
          0.2475839902489274,
          2.127184980007265,
          0.6570303412874574,
          0.16794730659465534,
          2.765258719399421,
          1.4382954540763537,
          1.5156718408966148,
          0.36408420706837397,
          0.7929241907202421,
          1.8946776149031732,
          0.7961551182097963,
          1.119176156404052,
          1.446397381069858,
          0.8294177917911054,
          1.0801651433977786,
          18.38649265135158,
        ],
        [3, 4, 5]);
    testUnary(
        'exp',
        [
          0.3143407,   0.03632548,  0.5354084,   -0.5000897,  1.2028517,
          -1.2581364,  -1.5108215,  -1.2340564,  1.3860914,   -0.2944251,
          -1.5065757,  -0.4673513,  0.56616277,  0.77866685,  -0.01097398,
          1.0758846,   0.6035437,   0.36806744,  0.03906458,  -0.54385495,
          0.10609569,  -0.40644982, -1.2890846,  1.3825086,   0.51489764,
          1.6407244,   -0.67886734, -0.6556329,  1.0399923,   0.1484657,
          1.011217,    0.8451463,   0.75473833,  -2.0161264,  1.6406634,
          -0.01692923, -0.7986609,  0.97758174,  0.893054,    -0.01632686,
          -1.9721986,  -0.75843745, 0.42327842,  -0.08648382, -1.3960054,
          0.7547995,   -0.42002508, -1.784105,   1.0171342,   0.3634587,
          0.4158588,   -1.0103701,  -0.23202766, 0.6390487,   -0.22796124,
          0.11259284,  0.3690759,   -0.18703128, 0.07711394,  2.9116163,
        ],
        [
          1.3693561967375985,
          1.036993312152022,
          1.708145706528859,
          0.6064762563524844,
          3.3295984129224556,
          0.2841831370156004,
          0.22072857493397907,
          0.29110932356722374,
          3.999188237901296,
          0.7449597417366962,
          0.2216677366530061,
          0.6266599061142515,
          1.7614948056979465,
          2.1785659734395244,
          0.9890860144586422,
          2.9325859189764807,
          1.828587297220383,
          1.4449394824067432,
          1.0398376341962599,
          0.580506111445723,
          1.111928271832297,
          0.666010515185213,
          0.27552288133235436,
          3.984885583357506,
          1.6734671969280184,
          5.158905269957428,
          0.5071911422650927,
          0.5191134117181747,
          2.829195229464421,
          1.1600530072738744,
          2.7489444455169916,
          2.328318422639379,
          2.127054864081674,
          0.13317031580877914,
          5.158590586333908,
          0.9832132641755142,
          0.4499310635787923,
          2.6580206785468157,
          2.4425779049391343,
          0.983805700764556,
          0.13915058318029658,
          0.4683977504013186,
          1.526959372817342,
          0.9171503881579351,
          0.2475839902489274,
          2.127184980007265,
          0.6570303412874574,
          0.16794730659465534,
          2.765258719399421,
          1.4382954540763537,
          1.5156718408966148,
          0.36408420706837397,
          0.7929241907202421,
          1.8946776149031732,
          0.7961551182097963,
          1.119176156404052,
          1.446397381069858,
          0.8294177917911054,
          1.0801651433977786,
          18.38649265135158,
        ],
        [3, 2, 2, 5]);
  }, 'exp');

  test(() => {
    testUnary('floor', [-1.1, 0, 1.1], [-2, 0, 1], [3]);
    testUnary(
        'floor',
        [-1.1, 0, 1.1, -2.2, 0, 2.2],
        [-2, 0, 1, -3, 0, 2],
        [2, 3]);
    testUnary(
        'floor',
        [-1.1, 0, 1.1, -2.2, 0, 2.2],
        [-2, 0, 1, -3, 0, 2],
        [1, 2, 3]);
    testUnary(
        'floor',
        [-1.1, 0, 1.1, -2.2, 0, 2.2],
        [-2, 0, 1, -3, 0, 2],
        [1, 2, 3, 1]);
  }, 'floor');

  test(() => {
    testUnary(
        'log',
        [1.4599811, 0.34325936, 1.0420732],
        [0.37842348, -1.069269, 0.04121224],
        [3]);
    testUnary(
        'log',
        [
          1.4599811,  0.34325936, 1.0420732, 0.10867598,
          0.39999306, 0.03704359, 1.5873954, 0.44784936,
          0.69070333, 1.8561625,  1.4088289, 0.06367786,
        ],
        [
          0.37842348,  -1.069269, 0.04121224, -2.2193844,
          -0.91630805, -3.29566,  0.4620946,  -0.80329835,
          -0.37004486, 0.6185112, 0.34275877, -2.7539184,
        ],
        [3, 4]);
    testUnary(
        'log',
        [
          1.4599811,  0.34325936, 1.0420732,  0.10867598, 0.39999306,
          0.03704359, 1.5873954,  0.44784936, 0.69070333, 1.8561625,
          1.4088289,  0.06367786, 0.32938832, 1.2429568,  1.1544572,
          0.47578564, 1.868428,   1.2279319,  1.0712656,  1.17982,
          1.460244,   0.62389,    0.79644215, 0.4196875,  0.372386,
          1.8887448,  1.4791015,  0.98091763, 0.45482925, 0.50871295,
          0.11605832, 0.86883324, 0.6235918,  1.392687,   0.75550365,
          0.35920736, 0.04935746, 0.13449927, 1.3587855,  0.9073937,
          1.0731584,  1.7933426,  1.9806778,  0.43379396, 1.3261564,
          0.52664477, 0.041302,   1.5167572,  0.6400343,  0.7669278,
          1.1766342,  1.6620969,  1.2579637,  1.7453014,  0.5470841,
          1.5960937,  0.37127188, 1.9055833,  1.3749765,  0.43101534,
        ],
        [
          0.37842348,  -1.069269,   0.04121224,  -2.2193844,  -0.91630805,
          -3.29566,    0.4620946,   -0.80329835, -0.37004486, 0.6185112,
          0.34275877,  -2.7539184,  -1.110518,   0.21749303,  0.1436303,
          -0.74278784, 0.62509745,  0.20533134,  0.06884073,  0.16536184,
          0.3786036,   -0.47178125, -0.22760078, -0.8682449,  -0.9878243,
          0.6359125,   0.39143485,  -0.01926679, -0.7878332,  -0.6758714,
          -2.1536624,  -0.14060406, -0.4722593,  0.33123496,  -0.28037068,
          -1.0238554,  -3.0086665,  -2.0061965,  0.3065913,   -0.09717887,
          0.07060606,  0.58408123,  0.68343914,  -0.83518565, 0.28228483,
          -0.64122903, -3.1868443,  0.4165747,   -0.44623348, -0.26536265,
          0.16265798,  0.50807995,  0.22949426,  0.55692726,  -0.60315275,
          0.46755916,  -0.99082065, 0.64478815,  0.31843665,  -0.8416116,
        ],
        [3, 4, 5]);
    testUnary(
        'log',
        [
          1.4599811,  0.34325936, 1.0420732,  0.10867598, 0.39999306,
          0.03704359, 1.5873954,  0.44784936, 0.69070333, 1.8561625,
          1.4088289,  0.06367786, 0.32938832, 1.2429568,  1.1544572,
          0.47578564, 1.868428,   1.2279319,  1.0712656,  1.17982,
          1.460244,   0.62389,    0.79644215, 0.4196875,  0.372386,
          1.8887448,  1.4791015,  0.98091763, 0.45482925, 0.50871295,
          0.11605832, 0.86883324, 0.6235918,  1.392687,   0.75550365,
          0.35920736, 0.04935746, 0.13449927, 1.3587855,  0.9073937,
          1.0731584,  1.7933426,  1.9806778,  0.43379396, 1.3261564,
          0.52664477, 0.041302,   1.5167572,  0.6400343,  0.7669278,
          1.1766342,  1.6620969,  1.2579637,  1.7453014,  0.5470841,
          1.5960937,  0.37127188, 1.9055833,  1.3749765,  0.43101534,
        ],
        [
          0.37842348,  -1.069269,   0.04121224,  -2.2193844,  -0.91630805,
          -3.29566,    0.4620946,   -0.80329835, -0.37004486, 0.6185112,
          0.34275877,  -2.7539184,  -1.110518,   0.21749303,  0.1436303,
          -0.74278784, 0.62509745,  0.20533134,  0.06884073,  0.16536184,
          0.3786036,   -0.47178125, -0.22760078, -0.8682449,  -0.9878243,
          0.6359125,   0.39143485,  -0.01926679, -0.7878332,  -0.6758714,
          -2.1536624,  -0.14060406, -0.4722593,  0.33123496,  -0.28037068,
          -1.0238554,  -3.0086665,  -2.0061965,  0.3065913,   -0.09717887,
          0.07060606,  0.58408123,  0.68343914,  -0.83518565, 0.28228483,
          -0.64122903, -3.1868443,  0.4165747,   -0.44623348, -0.26536265,
          0.16265798,  0.50807995,  0.22949426,  0.55692726,  -0.60315275,
          0.46755916,  -0.99082065, 0.64478815,  0.31843665,  -0.8416116,
        ],
        [3, 2, 2, 5]);
  }, 'log');

  test(() => {
    testUnary('neg', [-1.1, 0, 1.1], [1.1, 0, -1.1], [3]);
    testUnary(
        'neg',
        [-1, 0, 1.1, -2.2, 0, 2],
        [1, -0, -1.1, 2.2, -0, -2],
        [2, 3]);
    testUnary(
        'neg',
        [-1, 0, 1.1, -2.2, 0, 2],
        [1, -0, -1.1, 2.2, -0, -2],
        [1, 2, 3]);
    testUnary(
        'neg',
        [-1, 0, 1.1, -2.2, 0, 2],
        [1, -0, -1.1, 2.2, -0, -2],
        [1, 2, 3, 1]);
  }, 'neg');

  test(() => {
    testUnary(
        'sin',
        [1.4124068, 1.9740626, -0.06506752, 0.73539704],
        [
          0.9874825807196697,
          0.9197841363835013,
          -0.06502161610088251,
          0.6708816392565617,
        ],
        [4]);
    testUnary(
        'sin',
        [
          1.4124068,   1.9740626,  -0.06506752, 0.73539704,
          -0.56439203, 0.89806247, 0.12939146,  -0.34816208,
          -1.0759926,  0.66291636, 0.21504708,  -0.71527237,
        ],
        [
          0.9874825807196697,
          0.9197841363835013,
          -0.06502161610088251,
          0.6708816392565617,
          -0.5349022325592097,
          0.7821210521062475,
          0.12903071357901844,
          -0.34117073738540654,
          -0.8800619288707335,
          0.6154181431265636,
          0.21339342411295945,
          -0.6558230571220807,
        ],
        [3, 4]);
    testUnary(
        'sin',
        [
          1.4124068,   1.9740626,
          -0.06506752, 0.73539704,
          -0.56439203, 0.89806247,
          0.12939146,  -0.34816208,
          -1.0759926,  0.66291636,
          0.21504708,  -0.71527237,
        ],
        [
          0.9874825807196697,
          0.9197841363835013,
          -0.06502161610088251,
          0.6708816392565617,
          -0.5349022325592097,
          0.7821210521062475,
          0.12903071357901844,
          -0.34117073738540654,
          -0.8800619288707335,
          0.6154181431265636,
          0.21339342411295945,
          -0.6558230571220807,
        ],
        [3, 2, 2]);
    testUnary(
        'sin',
        [
          1.4124068,
          1.9740626,
          -0.06506752,
          0.73539704,
          -0.56439203,
          0.89806247,
          0.12939146,
          -0.34816208,
          -1.0759926,
          0.66291636,
          0.21504708,
          -0.71527237,
        ],
        [
          0.9874825807196697,
          0.9197841363835013,
          -0.06502161610088251,
          0.6708816392565617,
          -0.5349022325592097,
          0.7821210521062475,
          0.12903071357901844,
          -0.34117073738540654,
          -0.8800619288707335,
          0.6154181431265636,
          0.21339342411295945,
          -0.6558230571220807,
        ],
        [3, 2, 2, 1]);
  }, 'sin');

  test(() => {
    testUnary(
        'tan',
        [1.4124068, 1.9740626, -0.06506752, 0.73539704],
        [
          6.260663735197218,
          -2.3438487548949354,
          -0.06515950301265735,
          0.9046842034668978,
        ],
        [4]);
    testUnary(
        'tan',
        [
          1.4124068,   1.9740626,  -0.06506752, 0.73539704,
          -0.56439203, 0.89806247, 0.12939146,  -0.34816208,
          -1.0759926,  0.66291636, 0.21504708,  -0.71527237,
        ],
        [
          6.260663735197218,
          -2.3438487548949354,
          -0.06515950301265735,
          0.9046842034668978,
          -0.633084855036293,
          1.2551560992491184,
          0.1301184258508601,
          -0.3629470737734702,
          -1.8533123782006025,
          0.7807885239004154,
          0.2184245387683735,
          -0.8687380268391548,
        ],
        [3, 4]);
    testUnary(
        'tan',
        [
          1.4124068,   1.9740626,
          -0.06506752, 0.73539704,
          -0.56439203, 0.89806247,
          0.12939146,  -0.34816208,
          -1.0759926,  0.66291636,
          0.21504708,  -0.71527237,
        ],
        [
          6.260663735197218,
          -2.3438487548949354,
          -0.06515950301265735,
          0.9046842034668978,
          -0.633084855036293,
          1.2551560992491184,
          0.1301184258508601,
          -0.3629470737734702,
          -1.8533123782006025,
          0.7807885239004154,
          0.2184245387683735,
          -0.8687380268391548,
        ],
        [3, 2, 2]);
    testUnary(
        'tan',
        [
          1.4124068,
          1.9740626,
          -0.06506752,
          0.73539704,
          -0.56439203,
          0.89806247,
          0.12939146,
          -0.34816208,
          -1.0759926,
          0.66291636,
          0.21504708,
          -0.71527237,
        ],
        [
          6.260663735197218,
          -2.3438487548949354,
          -0.06515950301265735,
          0.9046842034668978,
          -0.633084855036293,
          1.2551560992491184,
          0.1301184258508601,
          -0.3629470737734702,
          -1.8533123782006025,
          0.7807885239004154,
          0.2184245387683735,
          -0.8687380268391548,
        ],
        [3, 2, 2, 1]);
  }, 'tan');
</script>
